feat(core): P1-7 validator closure items + P1-8 official JSON-Schema structural layer
P1-7:
- New document.source_descriptions_present rule enforces the spec's
  'at least one defined Source Description' requirement
- Root $self is now a known field in DocUnknownFieldRule (no longer
  flagged as unknown on conformant 1.1 documents)

P1-8:
- OfficialSchemaRule validates rawRoot against the official Arazzo
  JSON Schemas (bundled for 1.0/2025-10-15 and 1.1/2026-04-15), layered
  before semantic rules with schema.invalid error code
- Preprocessing: remote 2020-12 metaschema refs neutralized for offline
  use; sibling $refs hoisted into allOf (draft-7 style  handling in
  justinrainbow silently skipped required checks otherwise); $anchor
  local refs (#workflowId/#stepId) resolved to plain strings
- Errors capped at 25 and mapped onto the existing ErrorCollector API

Also corrects step timeout units to milliseconds per the 1.1 schema,
adds 1.1 targetSelectorType to Payload Replacements (jsonpointer/
jsonpath/xpath, applied via shared evaluators), and models 1.1 goto-
action parameters (parsed; applied by binding evaluated values into the
target workflow's input scope at goto dispatch).

+++ packages/core/src/Spec/Step.php
@@ -36,7 +36,7 @@ final readonly class Step
         public readonly ?bool $strictValidation = null,
         public readonly ?bool $idempotencyKey = null,
         public readonly ?string $idempotencyHeader = null,
-        public readonly ?float $timeout = null,
+        public readonly ?int $timeout = null, // duration in milliseconds
     ) {
     }
 }
